+++ src/Farmer/Arm/Sql.fs
@@ -5,7 +5,7 @@ open Farmer
 open Farmer.Sql
 open System.Net
 
-let servers = ResourceType("Microsoft.Sql/servers", "2019-06-01-preview")
+let servers = ResourceType("Microsoft.Sql/servers", "2022-05-01-preview")
 
 let elasticPools =
     ResourceType("Microsoft.Sql/servers/elasticPools", "2017-10-01-preview")
@@ -23,18 +23,111 @@ type DbKind =
     | Standalone of DbPurchaseModel
     | Pool of ResourceName
 
+type ActiveDirectoryPrincipalType =
+    | User
+    | Group
+
+type ActiveDirectoryAdminSettings =
+    {
+        /// Ideally same as AD name
+        Login: string
+        /// Active Directory object id of user or group
+        Sid: string
+        PrincipalType: ActiveDirectoryPrincipalType
+        AdOnlyAuth: bool
+    }
+
+let (|MixedModeAuth|AdOnlyAuth|SqlOnlyAuth|) activeDirAdmin =
+    match activeDirAdmin with
+    | Some x when x.AdOnlyAuth -> AdOnlyAuth(x)
+    | Some x when not x.AdOnlyAuth -> MixedModeAuth(x)
+    | _ -> SqlOnlyAuth
+
+type SqlServerADAdminJsonProperties =
+    {
+        administratorType: string
+        principalType: string
+        login: string
+        sid: string
+        azureADOnlyAuthentication: bool
+    }
+
+type SqlServerJsonProperties =
+    {
+        version: string
+        minimalTlsVersion: string
+        administratorLogin: string
+        administratorLoginPassword: string
+        administrators: SqlServerADAdminJsonProperties
+    }
+
 type Server =
     {
         ServerName: SqlAccountName
         Location: Location
         Credentials: {| Username: string
                         Password: SecureParameter |}
+        ActiveDirectoryAdmin: ActiveDirectoryAdminSettings option
         MinTlsVersion: TlsVersion option
         Tags: Map<string, string>
     }
 
+    member private this.BuildSqlSeverPropertiesBase() : SqlServerJsonProperties =
+        {
+            version = "12.0"
+            minimalTlsVersion =
+                match this.MinTlsVersion with
+                | Some Tls10 -> "1.0"
+                | Some Tls11 -> "1.1"
+                | Some Tls12 -> "1.2"
+                | None -> null
+            administratorLogin = null
+            administratorLoginPassword = null
+            administrators = Unchecked.defaultof<SqlServerADAdminJsonProperties>
+        }
+
+    member private this.BuildSqlServerADOnlyAdmin(x: ActiveDirectoryAdminSettings) : SqlServerADAdminJsonProperties =
+        {
+            administratorType = "ActiveDirectory"
+            principalType =
+                match x.PrincipalType with
+                | Group -> "Group"
+                | User -> "User"
+            login = x.Login
+            sid = x.Sid
+            azureADOnlyAuthentication = true
+        }
+
+    member private this.BuildSqlServerPropertiesWithMixedModeAdministrator
+        (x: ActiveDirectoryAdminSettings)
+        : SqlServerJsonProperties =
+        { this.BuildSqlSeverPropertiesBase() with
+            administratorLogin = this.Credentials.Username
+            administratorLoginPassword = this.Credentials.Password.ArmExpression.Eval()
+            administrators =
+                { this.BuildSqlServerADOnlyAdmin(x) with
+                    azureADOnlyAuthentication = false
+                }
+        }
+
+    member private this.BuildSqlServerPropertiesWithADOnlyAdministrator
+        (x: ActiveDirectoryAdminSettings)
+        : SqlServerJsonProperties =
+        { this.BuildSqlSeverPropertiesBase() with
+            administrators = this.BuildSqlServerADOnlyAdmin(x)
+        }
+
+    member private this.BuildSqlServerPropertiesWithSqlOnlyAdministrator() : SqlServerJsonProperties =
+        { this.BuildSqlSeverPropertiesBase() with
+            administratorLogin = this.Credentials.Username
+            administratorLoginPassword = this.Credentials.Password.ArmExpression.Eval()
+        }
+
     interface IParameters with
-        member this.SecureParameters = [ this.Credentials.Password ]
+        member this.SecureParameters =
+            match this.ActiveDirectoryAdmin with
+            | Some (x) when x.AdOnlyAuth -> []
+            | _ -> [ this.Credentials.Password ]
 
     interface IArmResource with
         member this.ResourceId = servers.resourceId this.ServerName.ResourceName
@@ -46,17 +139,10 @@ type Server =
                    tags = (this.Tags |> Map.add "displayName" this.ServerName.ResourceName.Value)
                ) with
                 properties =
-                    {|
-                        administratorLogin = this.Credentials.Username
-                        administratorLoginPassword = this.Credentials.Password.ArmExpression.Eval()
-                        version = "12.0"
-                        minimalTlsVersion =
-                            match this.MinTlsVersion with
-                            | Some Tls10 -> "1.0"
-                            | Some Tls11 -> "1.1"
-                            | Some Tls12 -> "1.2"
-                            | None -> null
-                    |}
+                    match this.ActiveDirectoryAdmin with
+                    | MixedModeAuth x -> this.BuildSqlServerPropertiesWithMixedModeAdministrator(x)
+                    | AdOnlyAuth x -> this.BuildSqlServerPropertiesWithADOnlyAdministrator(x)
+                    | SqlOnlyAuth -> this.BuildSqlServerPropertiesWithSqlOnlyAdministrator()
             |}
 
 module Servers =