+++ src/Farmer/Common.fs
@@ -1170,6 +1170,8 @@ module Storage =
         private
         | StorageResourceName of ResourceName
 
+        static member internal Empty = StorageResourceName ResourceName.Empty
+
         static member Create name =
             [
                 nonEmptyLengthBetween 3 63